put distribution convenience functions back in terms of rvars, and also make Distribution class use rvar primarily.
The reason for the latter change in particular is that the polymorphism over m makes m
redundant - the instance can do nothing with it.  Also, RVarT Identity is trivially liftable, so there's no need to provide the more general type as the basis.  All it will do is make Distribution implementers mad when they realize they can't use the monad of their choice ;-)

The reason for the former change is more philosophical (and a bit pragmatic), and certainly open for debate.  Basically, it makes common code (at least, my common code) typecheck with fewer annotations.  For example, in a do-block the result monad is constrained, and the type of the value generated is also often constrained.  Thus the expression "sampleFrom src (uniform 0 100)" requires no type annotation.

It may help users of RVarT in such cases that I've also added a Lift instance to lift
RVar to (RVarT m).  If this is not convenient enough, I'm open to other suggestions.